+++ tritonbench/utils/loader.py
@@ -5,5 +5,5 @@ def load_library(library_path: str):
     import torch
 
     prefix, _delimiter, so_file = library_path.partition("/")
-    so_full_path = REPO_PATH.joinpath("utils", prefix, ".data", so_file).resolve()
+    so_full_path = REPO_PATH.joinpath("tools", prefix, ".data", so_file).resolve()
     torch.ops.load_library(str(so_full_path))
